+++ packages/code-du-travail-api/src/routes/items/search.es.ts
@@ -1,19 +1,21 @@
 export const getDocumentBody = ({
   url,
   source,
-  isAll,
+  ids,
 }: {
   url?: string;
   source?: string;
-  isAll: boolean;
-}) => {
+  ids?: string;
+}): any => {
   const filter: any[] = [{ term: { isPublished: true } }];
-  const size = isAll ? 200 : 1;
   if (url) {
     filter.push({ term: { url } });
   }
   if (source) {
     filter.push({ term: { source } });
+  }
+  if (ids) {
+    filter.push({ ids: { values: ids } });
   }
   return {
     query: {
@@ -21,6 +23,6 @@ export const getDocumentBody = ({
         filter,
       },
     },
-    size,
+    size: 200,
   };
 };
