+++ packages/code-du-travail-frontend/__tests__/information.test.tsx
@@ -15,12 +15,7 @@ const mockedProps: EditorialContentDataWrapper = {
   },
 };
 
-jest.mock(
-  "../src/information",
-  jest.fn(() => {
-    processToHtml: () => "myHtml";
-  })
-);
+jest.mock("../src/information/htmlProcess.service");
 
 describe("Information Page", () => {
   let renderResult: RenderResult;
