+++ packages/code-du-travail-frontend/pages/contribution/[slug].jsx
@@ -30,8 +30,10 @@ class PageContribution extends React.Component {
 
     const contentUrl = extractMdxContentUrl(markdown);
     if (contentUrl) {
-      const fetchContent = await fetch(`${API_URL}/items?url=${contentUrl}`);
-      const content = await fetchContent.json();
+      const fetchContent = await fetch(
+        `${API_URL}/items?url=${contentUrl}&all=true`
+      );
+      const [content] = await fetchContent.json();
       return { content, data };
     }
