+++ packages/code-du-travail-frontend/src/content/Components/index.ts
@@ -1,0 +1,1 @@
+export * from "./ContentList";
