+++ packages/code-du-travail-frontend/src/content/content.service.ts
@@ -1,0 +1,26 @@
+import getConfig from "next/config";
+import { SOURCES } from "@socialgouv/cdtn-sources";
+import { Content, ContentType, EditorialContentData } from "cdtn-types";
+import { getToolByIds } from "../outils";
+
+const {
+  publicRuntimeConfig: { API_URL },
+} = getConfig();
+
+export const getContentBySlug = async (slug: string): Promise<any> => {
+  const responseContainer = await fetch(
+    `${API_URL}/items/${SOURCES.EDITORIAL_CONTENT}/${slug}`
+  );
+  if (!responseContainer.ok) {
+    return { statusCode: responseContainer.status };
+  }
+  return await responseContainer.json();
+};
+
+export const getContentById = async (id: string): Promise<any> => {
+  const responseContainer = await fetch(`${API_URL}/items/${id}`);
+  if (!responseContainer.ok) {
+    return { statusCode: responseContainer.status };
+  }
+  return await responseContainer.json();
+};
