+++ packages/code-du-travail-frontend/src/content/index.ts
@@ -1,0 +1,2 @@
+export * from "./content.service";
+export * from "./Components";
