+++ packages/code-du-travail-frontend/src/information/Components/BlockMarkdown.tsx
@@ -1,0 +1,7 @@
+import { processToHtml } from "../htmlProcess.service";
+import React from "react";
+
+export const BlockMarkdown = ({ block }) => {
+  const { html: htmlMarkdown } = block;
+  return <React.Fragment>{processToHtml(htmlMarkdown) as any}</React.Fragment>;
+};
