+++ packages/code-du-travail-frontend/src/information/Components/index.ts
@@ -1,3 +1,3 @@
-export * from "./InfoLink";
-export * from "./ContentBlocks";
+export * from "./Link";
+export * from "./BlockList";
 export * from "./Contents";
