+++ packages/code-du-travail-frontend/src/information/htmlProcess.service.ts
@@ -1,6 +1,6 @@
 import unified from "unified";
 import React from "react";
-import { InfoLink } from "./Components";
+import { Link } from "../information/Components";
 import htmlToHtmlAst from "rehype-parse";
 import htmlAstToReact from "rehype-react";
 
@@ -11,7 +11,7 @@ const processor = unified()
   .use(htmlAstToReact, {
     Fragment: React.Fragment,
     components: {
-      a: InfoLink,
+      a: Link,
     },
     createElement: React.createElement,
   });
