+++ packages/code-du-travail-frontend/src/information/index.ts
@@ -1,2 +1,2 @@
 export * from "./information.service";
-export * from "./htmlProcess.service";
+export * from "./Components";
