+++ packages/code-du-travail-frontend/src/information/information.service.ts
@@ -1,21 +1,6 @@
-import getConfig from "next/config";
-import { SOURCES } from "@socialgouv/cdtn-sources";
 import { Content, ContentType, EditorialContentData } from "cdtn-types";
 import { getToolByIds } from "../outils";
-
-const {
-  publicRuntimeConfig: { API_URL },
-} = getConfig();
-
-export const getContentBySlug = async (slug: string): Promise<any> => {
-  const responseContainer = await fetch(
-    `${API_URL}/items/${SOURCES.EDITORIAL_CONTENT}/${slug}`
-  );
-  if (!responseContainer.ok) {
-    return { statusCode: responseContainer.status };
-  }
-  return await responseContainer.json();
-};
+import { getContentBySlug } from "../content";
 
 export const getContentBlockIds = (data: Content[]): string[] => {
   return data.reduce((idsAcc: string[], content) => {
